+++ fabric-item-api-v1/src/testmod/resources/fabric.mod.json
@@ -14,7 +14,8 @@
       "net.fabricmc.fabric.test.item.FabricItemSettingsTests",
       "net.fabricmc.fabric.test.item.ItemUpdateAnimationTest",
       "net.fabricmc.fabric.test.item.ModifyItemAttributeModifiersCallbackTest",
-      "net.fabricmc.fabric.test.item.ArmorKnockbackResistanceTest"
+      "net.fabricmc.fabric.test.item.ArmorKnockbackResistanceTest",
+      "net.fabricmc.fabric.test.item.FoodGameInitializer"
     ],
     "client": [
       "net.fabricmc.fabric.test.item.client.TooltipTests"
